+++ public/app/features/dashboard/dynamicDashboardSrv.js
@@ -91,7 +91,7 @@ function (angular, _) {
     // returns a new panel clone or reuses a clone from previous iteration
     this.repeatRow = function(row) {
       var variables = this.dashboard.templating.list;
-      var variable = _.findWhere(variables, {name: row.repeat.replace('$', '')});
+      var variable = _.findWhere(variables, {name: row.repeat});
       if (!variable) {
         return;
       }
@@ -105,6 +105,8 @@ function (angular, _) {
 
       _.each(selected, function(option, index) {
         copy = self.getRowClone(row, index);
+        copy.scopedVars = {};
+        copy.scopedVars[variable.name] = option;
 
         for (i = 0; i < copy.panels.length; i++) {
           panel = copy.panels[i];