+++ public/app/filters/all.js
@@ -57,7 +57,11 @@ define(['angular', 'jquery', 'lodash', 'moment'], function (angular, $, _, momen
 
   module.filter('interpolateTemplateVars', function(templateSrv) {
     function interpolateTemplateVars(text, scope) {
-      return templateSrv.replaceWithText(text, scope.panel.scopedVars);
+      if (scope.panel) {
+        return templateSrv.replaceWithText(text, scope.panel.scopedVars);
+      } else {
+        return templateSrv.replaceWithText(text, scope.row.scopedVars);
+      }
     }
 
     interpolateTemplateVars.$stateful = true;